MGMT-15177: Align feature support level to support platform as filterable feature
